// src/openapi/model/api-area-tree.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ApiAreaTree
 */
export interface ApiAreaTree {
    /**
     * 
     * @type {number}
     * @memberof ApiAreaTree
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ApiAreaTree
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof ApiAreaTree
     */
    position?: number;
    /**
     * 
     * @type {number}
     * @memberof ApiAreaTree
     */
    parentId?: number;
    /**
     * 
     * @type {Array<number>}
     * @memberof ApiAreaTree
     */
    childrenIds?: Array<number>;
}