// src/openapi/model/api-genre.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ApiGenre
 */
export interface ApiGenre {
    /**
     * 
     * @type {number}
     * @memberof ApiGenre
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ApiGenre
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof ApiGenre
     */
    position: number;
    /**
     * 
     * @type {number}
     * @memberof ApiGenre
     */
    parentId?: number;
    /**
     * 
     * @type {Array<number>}
     * @memberof ApiGenre
     */
    childrenIds?: Array<number>;
}